Enhance e2e tests for OVN VPC NAT Gateway and VIP
- Adjusted the wait duration in the OVN VPC NAT Gateway e2e tests to 2 seconds for checking finalizer addition and resource deletion, improving test responsiveness.
- Refactored conditional checks for subnet protocols in VIP e2e tests to use switch-case statements for better readability and maintainability.
- Added default cases in protocol checks to handle dual stack scenarios more clearly.
- Simplified the loop for waiting on VIP finalizer addition to enhance code clarity.
