+++ test/e2e/iptables-vpc-nat-gw/e2e_test.go
@@ -2,16 +2,13 @@ package ovn_eip
 
 import (
 	"context"
-	"errors"
 	"flag"
 	"fmt"
-	"strconv"
 	"strings"
 	"testing"
 	"time"
 
 	dockernetwork "github.com/moby/moby/api/types/network"
-	corev1 "k8s.io/api/core/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	clientset "k8s.io/client-go/kubernetes"
@@ -40,36 +37,6 @@ const (
 	externalSubnetProvider = "ovn-vpc-external-network.kube-system"
 )
 
-const (
-	iperf2Port = "20288"
-	skipIperf  = false
-)
-
-const (
-	eipLimit = iota*5 + 10
-	updatedEIPLimit
-	newEIPLimit
-	specificIPLimit
-	defaultNicLimit
-)
-
-type qosParams struct {
-	vpc1Name       string
-	vpc2Name       string
-	vpc1SubnetName string
-	vpc2SubnetName string
-	vpcNat1GwName  string
-	vpcNat2GwName  string
-	vpc1EIPName    string
-	vpc2EIPName    string
-	vpc1FIPName    string
-	vpc2FIPName    string
-	vpc1PodName    string
-	vpc2PodName    string
-	attachDefName  string
-	subnetProvider string
-}
-
 func setupNetworkAttachmentDefinition(
 	f *framework.Framework,
 	dockerExtNetNetwork *dockernetwork.Inspect,
@@ -85,7 +52,7 @@ func setupNetworkAttachmentDefinition(
 	ginkgo.By("Getting docker network " + dockerExtNetName)
 	network, err := docker.NetworkInspect(dockerExtNetName)
 	framework.ExpectNoError(err, "getting docker network "+dockerExtNetName)
-	ginkgo.By("Getting network attachment definition " + externalNetworkName)
+	ginkgo.By("Getting or creating network attachment definition " + externalNetworkName)
 	attachConf := fmt.Sprintf(`{
 		"cniVersion": "0.3.0",
 		"type": "macvlan",
@@ -97,9 +64,16 @@ func setupNetworkAttachmentDefinition(
 		  "provider": "%s"
 		}
 	  }`, nicName, provider)
-	attachNet := framework.MakeNetworkAttachmentDefinition(externalNetworkName, framework.KubeOvnNamespace, attachConf)
-	attachNetClient.Create(attachNet)
-	nad := attachNetClient.Get(externalNetworkName)
+
+	// Try to get existing NAD first
+	nad, err := attachNetClient.NetworkAttachmentDefinitionInterface.Get(context.TODO(), externalNetworkName, metav1.GetOptions{})
+	if err != nil && k8serrors.IsNotFound(err) {
+		// NAD doesn't exist, create it
+		attachNet := framework.MakeNetworkAttachmentDefinition(externalNetworkName, framework.KubeOvnNamespace, attachConf)
+		nad = attachNetClient.Create(attachNet)
+	} else {
+		framework.ExpectNoError(err, "getting network attachment definition "+externalNetworkName)
+	}
 
 	ginkgo.By("Got network attachment definition " + nad.Name)
 
@@ -138,8 +112,16 @@ func setupNetworkAttachmentDefinition(
 			excludeIPs = append(excludeIPs, container.IPv6Address.Addr().String())
 		}
 	}
-	macvlanSubnet := framework.MakeSubnet(externalNetworkName, "", strings.Join(cidr, ","), strings.Join(gateway, ","), "", provider, excludeIPs, nil, nil)
-	_ = subnetClient.CreateSync(macvlanSubnet)
+
+	// Check if subnet already exists
+	_, err = subnetClient.SubnetInterface.Get(context.TODO(), externalNetworkName, metav1.GetOptions{})
+	if err != nil && k8serrors.IsNotFound(err) {
+		// Subnet doesn't exist, create it
+		macvlanSubnet := framework.MakeSubnet(externalNetworkName, "", strings.Join(cidr, ","), strings.Join(gateway, ","), "", provider, excludeIPs, nil, nil)
+		_ = subnetClient.CreateSync(macvlanSubnet)
+	} else {
+		framework.ExpectNoError(err, "getting subnet "+externalNetworkName)
+	}
 }
 
 func setupVpcNatGwTestEnvironment(
@@ -208,6 +190,54 @@ func cleanVpcNatGwTestEnvironment(
 	vpcClient.DeleteSync(vpcName)
 }
 
+// waitForIptablesEIPReady waits for an IptablesEIP to be ready
+func waitForIptablesEIPReady(eipClient *framework.IptablesEIPClient, eipName string, timeout time.Duration) *apiv1.IptablesEIP {
+	ginkgo.GinkgoHelper()
+	var eip *apiv1.IptablesEIP
+	deadline := time.Now().Add(timeout)
+	for time.Now().Before(deadline) {
+		eip = eipClient.Get(eipName)
+		if eip != nil && eip.Status.IP != "" && eip.Status.Ready {
+			framework.Logf("IptablesEIP %s is ready with IP: %s", eipName, eip.Status.IP)
+			return eip
+		}
+		time.Sleep(2 * time.Second)
+	}
+	framework.Failf("Timeout waiting for IptablesEIP %s to be ready", eipName)
+	return nil
+}
+
+// verifySubnetStatusAfterEIPOperation verifies subnet status after EIP operation
+func verifySubnetStatusAfterEIPOperation(subnetClient *framework.SubnetClient, subnetName string,
+	protocol, operation, shouldContainIP string,
+) {
+	ginkgo.GinkgoHelper()
+
+	subnet := subnetClient.Get(subnetName)
+	framework.Logf("Verifying subnet %s status after %s: Protocol=%s", subnetName, operation, protocol)
+
+	switch protocol {
+	case apiv1.ProtocolIPv4:
+		framework.Logf("V4 Status: Available=%.0f, Using=%.0f",
+			subnet.Status.V4AvailableIPs, subnet.Status.V4UsingIPs)
+		if shouldContainIP != "" {
+			framework.ExpectTrue(strings.Contains(subnet.Status.V4UsingIPRange, shouldContainIP),
+				"IP %s should be in V4UsingIPRange after %s", shouldContainIP, operation)
+		}
+	case apiv1.ProtocolIPv6:
+		framework.Logf("V6 Status: Available=%.0f, Using=%.0f",
+			subnet.Status.V6AvailableIPs, subnet.Status.V6UsingIPs)
+		if shouldContainIP != "" {
+			framework.ExpectTrue(strings.Contains(subnet.Status.V6UsingIPRange, shouldContainIP),
+				"IP %s should be in V6UsingIPRange after %s", shouldContainIP, operation)
+		}
+	case apiv1.ProtocolDual:
+		framework.Logf("Dual Stack Status: V4Available=%.0f, V4Using=%.0f, V6Available=%.0f, V6Using=%.0f",
+			subnet.Status.V4AvailableIPs, subnet.Status.V4UsingIPs,
+			subnet.Status.V6AvailableIPs, subnet.Status.V6UsingIPs)
+	}
+}
+
 var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 	f := framework.NewDefaultFramework("iptables-vpc-nat-gw")
 
@@ -241,37 +271,38 @@ var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 	var net2VpcName string
 	var net2EipName string
 
-	vpcName = "vpc-" + framework.RandomSuffix()
-	vpcNatGwName = "gw-" + framework.RandomSuffix()
-
-	fipVipName = "fip-vip-" + framework.RandomSuffix()
-	fipEipName = "fip-eip-" + framework.RandomSuffix()
-	fipName = "fip-" + framework.RandomSuffix()
-
-	dnatVipName = "dnat-vip-" + framework.RandomSuffix()
-	dnatEipName = "dnat-eip-" + framework.RandomSuffix()
-	dnatName = "dnat-" + framework.RandomSuffix()
-
-	// sharing case
-	sharedVipName = "shared-vip-" + framework.RandomSuffix()
-	sharedEipName = "shared-eip-" + framework.RandomSuffix()
-	sharedEipDnatName = "shared-eip-dnat-" + framework.RandomSuffix()
-	sharedEipSnatName = "shared-eip-snat-" + framework.RandomSuffix()
-	sharedEipFipShouldOkName = "shared-eip-fip-should-ok-" + framework.RandomSuffix()
-	sharedEipFipShouldFailName = "shared-eip-fip-should-fail-" + framework.RandomSuffix()
-
-	snatEipName = "snat-eip-" + framework.RandomSuffix()
-	snatName = "snat-" + framework.RandomSuffix()
-	overlaySubnetName = "overlay-subnet-" + framework.RandomSuffix()
-
-	net2AttachDefName = "net2-ovn-vpc-external-network-" + framework.RandomSuffix()
-	net2SubnetProvider = fmt.Sprintf("%s.%s", net2AttachDefName, framework.KubeOvnNamespace)
-	net2OverlaySubnetName = "net2-overlay-subnet-" + framework.RandomSuffix()
-	net2VpcNatGwName = "net2-gw-" + framework.RandomSuffix()
-	net2VpcName = "net2-vpc-" + framework.RandomSuffix()
-	net2EipName = "net2-eip-" + framework.RandomSuffix()
-
 	ginkgo.BeforeEach(func() {
+		randomSuffix := framework.RandomSuffix()
+		vpcName = "vpc-" + randomSuffix
+		vpcNatGwName = "gw-" + randomSuffix
+
+		fipVipName = "fip-vip-" + randomSuffix
+		fipEipName = "fip-eip-" + randomSuffix
+		fipName = "fip-" + randomSuffix
+
+		dnatVipName = "dnat-vip-" + randomSuffix
+		dnatEipName = "dnat-eip-" + randomSuffix
+		dnatName = "dnat-" + randomSuffix
+
+		// sharing case
+		sharedVipName = "shared-vip-" + randomSuffix
+		sharedEipName = "shared-eip-" + randomSuffix
+		sharedEipDnatName = "shared-eip-dnat-" + randomSuffix
+		sharedEipSnatName = "shared-eip-snat-" + randomSuffix
+		sharedEipFipShouldOkName = "shared-eip-fip-should-ok-" + randomSuffix
+		sharedEipFipShouldFailName = "shared-eip-fip-should-fail-" + randomSuffix
+
+		snatEipName = "snat-eip-" + randomSuffix
+		snatName = "snat-" + randomSuffix
+		overlaySubnetName = "overlay-subnet-" + randomSuffix
+
+		net2AttachDefName = "net2-ovn-vpc-external-network-" + randomSuffix
+		net2SubnetProvider = fmt.Sprintf("%s.%s", net2AttachDefName, framework.KubeOvnNamespace)
+		net2OverlaySubnetName = "net2-overlay-subnet-" + randomSuffix
+		net2VpcNatGwName = "net2-gw-" + randomSuffix
+		net2VpcName = "net2-vpc-" + randomSuffix
+		net2EipName = "net2-eip-" + randomSuffix
+
 		cs = f.ClientSet
 		attachNetClient = f.NetworkAttachmentDefinitionClientNS(framework.KubeOvnNamespace)
 		subnetClient = f.SubnetClient()
@@ -521,10 +552,10 @@ var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 		iptablesFIPClient.DeleteSync(sharedEipFipShouldOkName)
 		ginkgo.By("Deleting share iptables fip " + sharedEipFipShouldFailName)
 		iptablesFIPClient.DeleteSync(sharedEipFipShouldFailName)
-		ginkgo.By("Deleting share iptables dnat " + dnatName)
-		iptablesDnatRuleClient.DeleteSync(dnatName)
-		ginkgo.By("Deleting share iptables snat " + snatName)
-		iptablesSnatRuleClient.DeleteSync(snatName)
+		ginkgo.By("Deleting share iptables dnat " + sharedEipDnatName)
+		iptablesDnatRuleClient.DeleteSync(sharedEipDnatName)
+		ginkgo.By("Deleting share iptables snat " + sharedEipSnatName)
+		iptablesSnatRuleClient.DeleteSync(sharedEipSnatName)
 
 		ginkgo.By("Deleting iptables fip " + fipName)
 		iptablesFIPClient.DeleteSync(fipName)
@@ -618,54 +649,6 @@ var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 		vpcClient.DeleteSync(net2VpcName)
 	})
 
-	// Helper function to wait for EIP to be ready
-	_ = func(eipClient *framework.IptablesEIPClient, eipName string, timeout time.Duration) *apiv1.IptablesEIP {
-		ginkgo.GinkgoHelper()
-		var eip *apiv1.IptablesEIP
-		deadline := time.Now().Add(timeout)
-		for time.Now().Before(deadline) {
-			eip = eipClient.Get(eipName)
-			if eip != nil && eip.Status.IP != "" && eip.Status.Ready {
-				framework.Logf("IptablesEIP %s is ready with IP: %s", eipName, eip.Status.IP)
-				return eip
-			}
-			time.Sleep(2 * time.Second)
-		}
-		framework.Failf("Timeout waiting for IptablesEIP %s to be ready", eipName)
-		return nil
-	}
-
-	// Helper function to verify subnet status after EIP operation
-	_ = func(subnetClient *framework.SubnetClient, subnetName string,
-		protocol string, expectedAvailableDelta, expectedUsingDelta float64,
-		operation string, shouldContainIP string) {
-		ginkgo.GinkgoHelper()
-
-		subnet := subnetClient.Get(subnetName)
-		framework.Logf("Verifying subnet %s status after %s: Protocol=%s", subnetName, operation, protocol)
-
-		switch protocol {
-		case apiv1.ProtocolIPv4:
-			framework.Logf("V4 Status: Available=%.0f, Using=%.0f",
-				subnet.Status.V4AvailableIPs, subnet.Status.V4UsingIPs)
-			if shouldContainIP != "" {
-				framework.ExpectTrue(strings.Contains(subnet.Status.V4UsingIPRange, shouldContainIP),
-					"IP %s should be in V4UsingIPRange after %s", shouldContainIP, operation)
-			}
-		case apiv1.ProtocolIPv6:
-			framework.Logf("V6 Status: Available=%.0f, Using=%.0f",
-				subnet.Status.V6AvailableIPs, subnet.Status.V6UsingIPs)
-			if shouldContainIP != "" {
-				framework.ExpectTrue(strings.Contains(subnet.Status.V6UsingIPRange, shouldContainIP),
-					"IP %s should be in V6UsingIPRange after %s", shouldContainIP, operation)
-			}
-		case apiv1.ProtocolDual:
-			framework.Logf("Dual Stack Status: V4Available=%.0f, V4Using=%.0f, V6Available=%.0f, V6Using=%.0f",
-				subnet.Status.V4AvailableIPs, subnet.Status.V4UsingIPs,
-				subnet.Status.V6AvailableIPs, subnet.Status.V6UsingIPs)
-		}
-	}
-
 	framework.ConformanceIt("should properly manage IptablesEIP lifecycle with finalizer and update subnet status", func() {
 		f.SkipVersionPriorTo(1, 13, "This feature was introduced in v1.13")
 
@@ -700,21 +683,13 @@ var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 		eip := framework.MakeIptablesEIP(eipName, "", "", "", vpcNatGwName, "", "")
 		_ = iptablesEIPClient.CreateSync(eip)
 
-		ginkgo.By("3. Wait for IptablesEIP CR to be created and get IP")
-		var eipCR *apiv1.IptablesEIP
-		for i := 0; i < 60; i++ {
-			eipCR = iptablesEIPClient.Get(eipName)
-			if eipCR != nil && eipCR.Status.IP != "" && eipCR.Status.Ready {
-				break
-			}
-			time.Sleep(1 * time.Second)
-		}
-		framework.ExpectNotNil(eipCR, "IptablesEIP CR should be created")
+		ginkgo.By("3. Wait for IptablesEIP CR to be ready")
+		eipCR := waitForIptablesEIPReady(iptablesEIPClient, eipName, 60*time.Second)
+		framework.ExpectNotNil(eipCR, "IptablesEIP CR should be created and ready")
 		framework.ExpectNotEmpty(eipCR.Status.IP, "IptablesEIP should have IP assigned")
-		framework.ExpectTrue(eipCR.Status.Ready, "IptablesEIP should be ready")
 
 		ginkgo.By("4. Wait for IptablesEIP CR finalizer to be added")
-		for i := 0; i < 60; i++ {
+		for range 60 {
 			eipCR = iptablesEIPClient.Get(eipName)
 			if eipCR != nil && len(eipCR.Finalizers) > 0 {
 				break
@@ -730,41 +705,30 @@ var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 
 		ginkgo.By("6. Verify external subnet status after IptablesEIP CR creation")
 		afterCreateSubnet := subnetClient.Get(externalSubnetName)
-		if afterCreateSubnet.Spec.Protocol == apiv1.ProtocolIPv4 {
-			// Verify IP count changed
+		verifySubnetStatusAfterEIPOperation(subnetClient, externalSubnetName,
+			afterCreateSubnet.Spec.Protocol, "IptablesEIP creation", eipCR.Status.IP)
+
+		// Verify IP count and range changes
+		switch afterCreateSubnet.Spec.Protocol {
+		case apiv1.ProtocolIPv4:
 			framework.ExpectEqual(initialV4AvailableIPs-1, afterCreateSubnet.Status.V4AvailableIPs,
 				"V4AvailableIPs should decrease by 1 after IptablesEIP creation")
 			framework.ExpectEqual(initialV4UsingIPs+1, afterCreateSubnet.Status.V4UsingIPs,
 				"V4UsingIPs should increase by 1 after IptablesEIP creation")
-
-			// Verify IP range changed
 			framework.ExpectNotEqual(initialV4AvailableIPRange, afterCreateSubnet.Status.V4AvailableIPRange,
 				"V4AvailableIPRange should change after IptablesEIP creation")
 			framework.ExpectNotEqual(initialV4UsingIPRange, afterCreateSubnet.Status.V4UsingIPRange,
 				"V4UsingIPRange should change after IptablesEIP creation")
-
-			// Verify the EIP's address is in the using range
-			eipIP := eipCR.Status.IP
-			framework.ExpectTrue(strings.Contains(afterCreateSubnet.Status.V4UsingIPRange, eipIP),
-				"EIP IP %s should be in V4UsingIPRange %s", eipIP, afterCreateSubnet.Status.V4UsingIPRange)
-		} else if afterCreateSubnet.Spec.Protocol == apiv1.ProtocolIPv6 {
-			// Verify IP count changed
+		case apiv1.ProtocolIPv6:
 			framework.ExpectEqual(initialV6AvailableIPs-1, afterCreateSubnet.Status.V6AvailableIPs,
 				"V6AvailableIPs should decrease by 1 after IptablesEIP creation")
 			framework.ExpectEqual(initialV6UsingIPs+1, afterCreateSubnet.Status.V6UsingIPs,
 				"V6UsingIPs should increase by 1 after IptablesEIP creation")
-
-			// Verify IP range changed
 			framework.ExpectNotEqual(initialV6AvailableIPRange, afterCreateSubnet.Status.V6AvailableIPRange,
 				"V6AvailableIPRange should change after IptablesEIP creation")
 			framework.ExpectNotEqual(initialV6UsingIPRange, afterCreateSubnet.Status.V6UsingIPRange,
 				"V6UsingIPRange should change after IptablesEIP creation")
-
-			// Verify the EIP's address is in the using range
-			eipIP := eipCR.Status.IP
-			framework.ExpectTrue(strings.Contains(afterCreateSubnet.Status.V6UsingIPRange, eipIP),
-				"EIP IP %s should be in V6UsingIPRange %s", eipIP, afterCreateSubnet.Status.V6UsingIPRange)
-		} else {
+		default:
 			// Dual stack
 			framework.ExpectEqual(initialV4AvailableIPs-1, afterCreateSubnet.Status.V4AvailableIPs,
 				"V4AvailableIPs should decrease by 1 after IptablesEIP creation")
@@ -774,7 +738,6 @@ var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 				"V6AvailableIPs should decrease by 1 after IptablesEIP creation")
 			framework.ExpectEqual(initialV6UsingIPs+1, afterCreateSubnet.Status.V6UsingIPs,
 				"V6UsingIPs should increase by 1 after IptablesEIP creation")
-
 			framework.ExpectNotEqual(initialV4AvailableIPRange, afterCreateSubnet.Status.V4AvailableIPRange,
 				"V4AvailableIPRange should change after IptablesEIP creation")
 			framework.ExpectNotEqual(initialV4UsingIPRange, afterCreateSubnet.Status.V4UsingIPRange,
@@ -800,7 +763,7 @@ var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 
 		ginkgo.By("8. Wait for IptablesEIP CR to be deleted")
 		deleted := false
-		for i := 0; i < 30; i++ {
+		for range 30 {
 			_, err := f.KubeOVNClientSet.KubeovnV1().IptablesEIPs().Get(context.Background(), eipName, metav1.GetOptions{})
 			if err != nil && k8serrors.IsNotFound(err) {
 				deleted = true
@@ -815,7 +778,12 @@ var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 
 		ginkgo.By("10. Verify external subnet status after IptablesEIP CR deletion")
 		afterDeleteSubnet := subnetClient.Get(externalSubnetName)
-		if afterDeleteSubnet.Spec.Protocol == apiv1.ProtocolIPv4 {
+		verifySubnetStatusAfterEIPOperation(subnetClient, externalSubnetName,
+			afterDeleteSubnet.Spec.Protocol, "IptablesEIP deletion", "")
+
+		// Verify IP count and range restoration
+		switch afterDeleteSubnet.Spec.Protocol {
+		case apiv1.ProtocolIPv4:
 			// Verify IP count is restored
 			framework.ExpectEqual(afterCreateV4AvailableIPs+1, afterDeleteSubnet.Status.V4AvailableIPs,
 				"V4AvailableIPs should increase by 1 after IptablesEIP deletion")
@@ -833,7 +801,7 @@ var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 				"V4AvailableIPs should return to initial value after IptablesEIP deletion")
 			framework.ExpectEqual(initialV4UsingIPs, afterDeleteSubnet.Status.V4UsingIPs,
 				"V4UsingIPs should return to initial value after IptablesEIP deletion")
-		} else if afterDeleteSubnet.Spec.Protocol == apiv1.ProtocolIPv6 {
+		case apiv1.ProtocolIPv6:
 			// Verify IP count is restored
 			framework.ExpectEqual(afterCreateV6AvailableIPs+1, afterDeleteSubnet.Status.V6AvailableIPs,
 				"V6AvailableIPs should increase by 1 after IptablesEIP deletion")
@@ -851,7 +819,7 @@ var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 				"V6AvailableIPs should return to initial value after IptablesEIP deletion")
 			framework.ExpectEqual(initialV6UsingIPs, afterDeleteSubnet.Status.V6UsingIPs,
 				"V6UsingIPs should return to initial value after IptablesEIP deletion")
-		} else {
+		default:
 			// Dual stack
 			framework.ExpectEqual(afterCreateV4AvailableIPs+1, afterDeleteSubnet.Status.V4AvailableIPs,
 				"V4AvailableIPs should increase by 1 after IptablesEIP deletion")
@@ -913,16 +881,8 @@ var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 		_ = iptablesEIPClient.CreateSync(eip)
 
 		ginkgo.By("3. Wait for IptablesEIP to be ready")
-		var eipCR *apiv1.IptablesEIP
-		for i := 0; i < 60; i++ {
-			eipCR = iptablesEIPClient.Get(eipName)
-			if eipCR != nil && eipCR.Status.IP != "" && eipCR.Status.Ready {
-				break
-			}
-			time.Sleep(1 * time.Second)
-		}
-		framework.ExpectNotNil(eipCR, "IptablesEIP CR should be created")
-		framework.ExpectTrue(eipCR.Status.Ready, "IptablesEIP should be ready")
+		eipCR := waitForIptablesEIPReady(iptablesEIPClient, eipName, 60*time.Second)
+		framework.ExpectNotNil(eipCR, "IptablesEIP CR should be created and ready")
 
 		ginkgo.By("4. Create IptablesFIP using the EIP")
 		fipName := "test-fip-" + framework.RandomSuffix()
@@ -930,7 +890,7 @@ var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 		_ = iptablesFIPClient.CreateSync(fip)
 
 		ginkgo.By("5. Wait for EIP status to show it's being used by FIP")
-		for i := 0; i < 60; i++ {
+		for range 60 {
 			eipCR = iptablesEIPClient.Get(eipName)
 			if eipCR != nil && strings.Contains(eipCR.Status.Nat, util.FipUsingEip) {
 				break
@@ -957,7 +917,7 @@ var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 
 		ginkgo.By("9. Wait for FIP to be deleted")
 		fipDeleted := false
-		for i := 0; i < 30; i++ {
+		for range 30 {
 			_, err := f.KubeOVNClientSet.KubeovnV1().IptablesFIPRules().Get(context.Background(), fipName, metav1.GetOptions{})
 			if err != nil && k8serrors.IsNotFound(err) {
 				fipDeleted = true
@@ -967,10 +927,15 @@ var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 		}
 		framework.ExpectTrue(fipDeleted, "FIP should be deleted")
 
-		ginkgo.By("10. Wait for EIP status.Nat to be cleared")
-		for i := 0; i < 30; i++ {
-			eipCR = iptablesEIPClient.Get(eipName)
-			if eipCR == nil || eipCR.Status.Nat == "" {
+		ginkgo.By("10. Wait for EIP status.Nat to be cleared or EIP to be deleted")
+		for range 30 {
+			eipCR, err := f.KubeOVNClientSet.KubeovnV1().IptablesEIPs().Get(context.Background(), eipName, metav1.GetOptions{})
+			if err != nil && k8serrors.IsNotFound(err) {
+				// EIP already deleted, which is expected
+				break
+			}
+			framework.ExpectNoError(err, "Failed to get IptablesEIP")
+			if eipCR.Status.Nat == "" {
 				break
 			}
 			time.Sleep(1 * time.Second)
@@ -978,7 +943,7 @@ var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 
 		ginkgo.By("11. Verify EIP is now deleted after FIP is removed")
 		eipDeleted := false
-		for i := 0; i < 30; i++ {
+		for range 30 {
 			_, err := f.KubeOVNClientSet.KubeovnV1().IptablesEIPs().Get(context.Background(), eipName, metav1.GetOptions{})
 			if err != nil && k8serrors.IsNotFound(err) {
 				eipDeleted = true
@@ -992,780 +957,6 @@ var _ = framework.SerialDescribe("[group:iptables-vpc-nat-gw]", func() {
 		vipClient.DeleteSync(vipName)
 
 		ginkgo.By("13. Test completed: IptablesEIP finalizer correctly blocks deletion when used by NAT rules")
-	})
-})
-
-func iperf(f *framework.Framework, iperfClientPod *corev1.Pod, iperfServerEIP *apiv1.IptablesEIP) string {
-	ginkgo.GinkgoHelper()
-
-	for i := range 20 {
-		command := fmt.Sprintf("iperf -e -p %s --reportstyle C -i 1 -c %s -t 10", iperf2Port, iperfServerEIP.Status.IP)
-		stdOutput, errOutput, err := framework.ExecShellInPod(context.Background(), f, iperfClientPod.Namespace, iperfClientPod.Name, command)
-		framework.Logf("output from exec on client pod %s (eip %s)\n", iperfClientPod.Name, iperfServerEIP.Name)
-		if stdOutput != "" && err == nil {
-			framework.Logf("output:\n%s", stdOutput)
-			return stdOutput
-		}
-		framework.Logf("exec %s failed err: %v, errOutput: %s, stdOutput: %s, retried %d times.", command, err, errOutput, stdOutput, i)
-		time.Sleep(6 * time.Second)
-	}
-	framework.ExpectNoError(errors.New("iperf failed"))
-	return ""
-}
-
-func checkQos(f *framework.Framework,
-	vpc1Pod, vpc2Pod *corev1.Pod, vpc1EIP, vpc2EIP *apiv1.IptablesEIP,
-	limit int, expect bool,
-) {
-	ginkgo.GinkgoHelper()
-
-	if !skipIperf {
-		if expect {
-			output := iperf(f, vpc1Pod, vpc2EIP)
-			framework.ExpectTrue(validRateLimit(output, limit))
-			output = iperf(f, vpc2Pod, vpc1EIP)
-			framework.ExpectTrue(validRateLimit(output, limit))
-		} else {
-			output := iperf(f, vpc1Pod, vpc2EIP)
-			framework.ExpectFalse(validRateLimit(output, limit))
-			output = iperf(f, vpc2Pod, vpc1EIP)
-			framework.ExpectFalse(validRateLimit(output, limit))
-		}
-	}
-}
-
-func newVPCQoSParamsInit() *qosParams {
-	qosParams := &qosParams{
-		vpc1Name:       "qos-vpc1-" + framework.RandomSuffix(),
-		vpc2Name:       "qos-vpc2-" + framework.RandomSuffix(),
-		vpc1SubnetName: "qos-vpc1-subnet-" + framework.RandomSuffix(),
-		vpc2SubnetName: "qos-vpc2-subnet-" + framework.RandomSuffix(),
-		vpcNat1GwName:  "qos-vpc1-gw-" + framework.RandomSuffix(),
-		vpcNat2GwName:  "qos-vpc2-gw-" + framework.RandomSuffix(),
-		vpc1EIPName:    "qos-vpc1-eip-" + framework.RandomSuffix(),
-		vpc2EIPName:    "qos-vpc2-eip-" + framework.RandomSuffix(),
-		vpc1FIPName:    "qos-vpc1-fip-" + framework.RandomSuffix(),
-		vpc2FIPName:    "qos-vpc2-fip-" + framework.RandomSuffix(),
-		vpc1PodName:    "qos-vpc1-pod-" + framework.RandomSuffix(),
-		vpc2PodName:    "qos-vpc2-pod-" + framework.RandomSuffix(),
-		attachDefName:  "qos-ovn-vpc-external-network-" + framework.RandomSuffix(),
-	}
-	qosParams.subnetProvider = fmt.Sprintf("%s.%s", qosParams.attachDefName, framework.KubeOvnNamespace)
-	return qosParams
-}
-
-func getNicDefaultQoSPolicy(limit int) apiv1.QoSPolicyBandwidthLimitRules {
-	return apiv1.QoSPolicyBandwidthLimitRules{
-		apiv1.QoSPolicyBandwidthLimitRule{
-			Name:      "net1-ingress",
-			Interface: "net1",
-			RateMax:   strconv.Itoa(limit),
-			BurstMax:  strconv.Itoa(limit),
-			Priority:  3,
-			Direction: apiv1.QoSDirectionIngress,
-		},
-		apiv1.QoSPolicyBandwidthLimitRule{
-			Name:      "net1-egress",
-			Interface: "net1",
-			RateMax:   strconv.Itoa(limit),
-			BurstMax:  strconv.Itoa(limit),
-			Priority:  3,
-			Direction: apiv1.QoSDirectionEgress,
-		},
-	}
-}
-
-func getEIPQoSRule(limit int) apiv1.QoSPolicyBandwidthLimitRules {
-	return apiv1.QoSPolicyBandwidthLimitRules{
-		apiv1.QoSPolicyBandwidthLimitRule{
-			Name:      "eip-ingress",
-			RateMax:   strconv.Itoa(limit),
-			BurstMax:  strconv.Itoa(limit),
-			Priority:  1,
-			Direction: apiv1.QoSDirectionIngress,
-		},
-		apiv1.QoSPolicyBandwidthLimitRule{
-			Name:      "eip-egress",
-			RateMax:   strconv.Itoa(limit),
-			BurstMax:  strconv.Itoa(limit),
-			Priority:  1,
-			Direction: apiv1.QoSDirectionEgress,
-		},
-	}
-}
-
-func getSpecialQoSRule(limit int, ip string) apiv1.QoSPolicyBandwidthLimitRules {
-	return apiv1.QoSPolicyBandwidthLimitRules{
-		apiv1.QoSPolicyBandwidthLimitRule{
-			Name:       "net1-extip-ingress",
-			Interface:  "net1",
-			RateMax:    strconv.Itoa(limit),
-			BurstMax:   strconv.Itoa(limit),
-			Priority:   2,
-			Direction:  apiv1.QoSDirectionIngress,
-			MatchType:  apiv1.QoSMatchTypeIP,
-			MatchValue: "src " + ip + "/32",
-		},
-		apiv1.QoSPolicyBandwidthLimitRule{
-			Name:       "net1-extip-egress",
-			Interface:  "net1",
-			RateMax:    strconv.Itoa(limit),
-			BurstMax:   strconv.Itoa(limit),
-			Priority:   2,
-			Direction:  apiv1.QoSDirectionEgress,
-			MatchType:  apiv1.QoSMatchTypeIP,
-			MatchValue: "dst " + ip + "/32",
-		},
-	}
-}
-
-// defaultQoSCases test default qos policy=
-func defaultQoSCases(f *framework.Framework,
-	vpcNatGwClient *framework.VpcNatGatewayClient,
-	podClient *framework.PodClient,
-	qosPolicyClient *framework.QoSPolicyClient,
-	vpc1Pod *corev1.Pod,
-	vpc2Pod *corev1.Pod,
-	vpc1EIP *apiv1.IptablesEIP,
-	vpc2EIP *apiv1.IptablesEIP,
-	natgwName string,
-) {
-	ginkgo.GinkgoHelper()
-
-	// create nic qos policy
-	qosPolicyName := "default-nic-qos-policy-" + framework.RandomSuffix()
-	ginkgo.By("Creating qos policy " + qosPolicyName)
-	rules := getNicDefaultQoSPolicy(defaultNicLimit)
-
-	qosPolicy := framework.MakeQoSPolicy(qosPolicyName, true, apiv1.QoSBindingTypeNatGw, rules)
-	_ = qosPolicyClient.CreateSync(qosPolicy)
-
-	ginkgo.By("Patch natgw " + natgwName + " with qos policy " + qosPolicyName)
-	_ = vpcNatGwClient.PatchQoSPolicySync(natgwName, qosPolicyName)
-
-	ginkgo.By("Check qos " + qosPolicyName + " is limited to " + strconv.Itoa(defaultNicLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, defaultNicLimit, true)
-
-	ginkgo.By("Delete natgw pod " + natgwName + "-0")
-	natGwPodName := util.GenNatGwPodName(natgwName)
-	podClient.DeleteSync(natGwPodName)
-
-	ginkgo.By("Wait for natgw " + natgwName + "qos rebuild")
-	time.Sleep(5 * time.Second)
-
-	ginkgo.By("Check qos " + qosPolicyName + " is limited to " + strconv.Itoa(defaultNicLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, defaultNicLimit, true)
-
-	ginkgo.By("Remove qos policy " + qosPolicyName + " from natgw " + natgwName)
-	_ = vpcNatGwClient.PatchQoSPolicySync(natgwName, "")
-
-	ginkgo.By("Deleting qos policy " + qosPolicyName)
-	qosPolicyClient.DeleteSync(qosPolicyName)
-
-	ginkgo.By("Check qos " + qosPolicyName + " is not limited to " + strconv.Itoa(defaultNicLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, defaultNicLimit, false)
-}
-
-// eipQoSCases test default qos policy
-func eipQoSCases(f *framework.Framework,
-	eipClient *framework.IptablesEIPClient,
-	podClient *framework.PodClient,
-	qosPolicyClient *framework.QoSPolicyClient,
-	vpc1Pod *corev1.Pod,
-	vpc2Pod *corev1.Pod,
-	vpc1EIP *apiv1.IptablesEIP,
-	vpc2EIP *apiv1.IptablesEIP,
-	eipName string,
-	natgwName string,
-) {
-	ginkgo.GinkgoHelper()
-
-	// create eip qos policy
-	qosPolicyName := "eip-qos-policy-" + framework.RandomSuffix()
-	ginkgo.By("Creating qos policy " + qosPolicyName)
-	rules := getEIPQoSRule(eipLimit)
-
-	qosPolicy := framework.MakeQoSPolicy(qosPolicyName, false, apiv1.QoSBindingTypeEIP, rules)
-	qosPolicy = qosPolicyClient.CreateSync(qosPolicy)
-
-	ginkgo.By("Patch eip " + eipName + " with qos policy " + qosPolicyName)
-	_ = eipClient.PatchQoSPolicySync(eipName, qosPolicyName)
-
-	ginkgo.By("Check qos " + qosPolicyName + " is limited to " + strconv.Itoa(eipLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, eipLimit, true)
-
-	ginkgo.By("Update qos policy " + qosPolicyName + " with new rate limit")
-
-	rules = getEIPQoSRule(updatedEIPLimit)
-	modifiedqosPolicy := qosPolicy.DeepCopy()
-	modifiedqosPolicy.Spec.BandwidthLimitRules = rules
-	qosPolicyClient.Patch(qosPolicy, modifiedqosPolicy)
-	qosPolicyClient.WaitToQoSReady(qosPolicyName)
-
-	ginkgo.By("Check qos " + qosPolicyName + " is changed to " + strconv.Itoa(updatedEIPLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, updatedEIPLimit, true)
-
-	ginkgo.By("Delete natgw pod " + natgwName + "-0")
-	natGwPodName := util.GenNatGwPodName(natgwName)
-	podClient.DeleteSync(natGwPodName)
-
-	ginkgo.By("Wait for natgw " + natgwName + "qos rebuid")
-	time.Sleep(5 * time.Second)
-
-	ginkgo.By("Check qos " + qosPolicyName + " is limited to " + strconv.Itoa(updatedEIPLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, updatedEIPLimit, true)
-
-	newQoSPolicyName := "new-eip-qos-policy-" + framework.RandomSuffix()
-	newRules := getEIPQoSRule(newEIPLimit)
-	newQoSPolicy := framework.MakeQoSPolicy(newQoSPolicyName, false, apiv1.QoSBindingTypeEIP, newRules)
-	_ = qosPolicyClient.CreateSync(newQoSPolicy)
-
-	ginkgo.By("Change qos policy of eip " + eipName + " to " + newQoSPolicyName)
-	_ = eipClient.PatchQoSPolicySync(eipName, newQoSPolicyName)
-
-	ginkgo.By("Check qos " + qosPolicyName + " is limited to " + strconv.Itoa(newEIPLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, newEIPLimit, true)
-
-	ginkgo.By("Remove qos policy " + qosPolicyName + " from natgw " + eipName)
-	_ = eipClient.PatchQoSPolicySync(eipName, "")
-
-	ginkgo.By("Deleting qos policy " + qosPolicyName)
-	qosPolicyClient.DeleteSync(qosPolicyName)
-
-	ginkgo.By("Deleting qos policy " + newQoSPolicyName)
-	qosPolicyClient.DeleteSync(newQoSPolicyName)
-
-	ginkgo.By("Check qos " + qosPolicyName + " is not limited to " + strconv.Itoa(newEIPLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, newEIPLimit, false)
-}
-
-// specifyingIPQoSCases test default qos policy
-func specifyingIPQoSCases(f *framework.Framework,
-	vpcNatGwClient *framework.VpcNatGatewayClient,
-	qosPolicyClient *framework.QoSPolicyClient,
-	vpc1Pod *corev1.Pod,
-	vpc2Pod *corev1.Pod,
-	vpc1EIP *apiv1.IptablesEIP,
-	vpc2EIP *apiv1.IptablesEIP,
-	natgwName string,
-) {
-	ginkgo.GinkgoHelper()
-
-	// create nic qos policy
-	qosPolicyName := "specifying-ip-qos-policy-" + framework.RandomSuffix()
-	ginkgo.By("Creating qos policy " + qosPolicyName)
-
-	rules := getSpecialQoSRule(specificIPLimit, vpc2EIP.Status.IP)
-
-	qosPolicy := framework.MakeQoSPolicy(qosPolicyName, true, apiv1.QoSBindingTypeNatGw, rules)
-	_ = qosPolicyClient.CreateSync(qosPolicy)
-
-	ginkgo.By("Patch natgw " + natgwName + " with qos policy " + qosPolicyName)
-	_ = vpcNatGwClient.PatchQoSPolicySync(natgwName, qosPolicyName)
-
-	ginkgo.By("Check qos " + qosPolicyName + " is limited to " + strconv.Itoa(specificIPLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, specificIPLimit, true)
-
-	ginkgo.By("Remove qos policy " + qosPolicyName + " from natgw " + natgwName)
-	_ = vpcNatGwClient.PatchQoSPolicySync(natgwName, "")
-
-	ginkgo.By("Deleting qos policy " + qosPolicyName)
-	qosPolicyClient.DeleteSync(qosPolicyName)
-
-	ginkgo.By("Check qos " + qosPolicyName + " is not limited to " + strconv.Itoa(specificIPLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, specificIPLimit, false)
-}
-
-// priorityQoSCases test qos match priority
-func priorityQoSCases(f *framework.Framework,
-	vpcNatGwClient *framework.VpcNatGatewayClient,
-	eipClient *framework.IptablesEIPClient,
-	qosPolicyClient *framework.QoSPolicyClient,
-	vpc1Pod *corev1.Pod,
-	vpc2Pod *corev1.Pod,
-	vpc1EIP *apiv1.IptablesEIP,
-	vpc2EIP *apiv1.IptablesEIP,
-	natgwName string,
-	eipName string,
-) {
-	ginkgo.GinkgoHelper()
-
-	// create nic qos policy
-	natGwQoSPolicyName := "priority-nic-qos-policy-" + framework.RandomSuffix()
-	ginkgo.By("Creating qos policy " + natGwQoSPolicyName)
-	// default qos policy + special qos policy
-	natgwRules := getNicDefaultQoSPolicy(defaultNicLimit)
-	natgwRules = append(natgwRules, getSpecialQoSRule(specificIPLimit, vpc2EIP.Status.IP)...)
-
-	natgwQoSPolicy := framework.MakeQoSPolicy(natGwQoSPolicyName, true, apiv1.QoSBindingTypeNatGw, natgwRules)
-	_ = qosPolicyClient.CreateSync(natgwQoSPolicy)
-
-	ginkgo.By("Patch natgw " + natgwName + " with qos policy " + natGwQoSPolicyName)
-	_ = vpcNatGwClient.PatchQoSPolicySync(natgwName, natGwQoSPolicyName)
-
-	eipQoSPolicyName := "eip-qos-policy-" + framework.RandomSuffix()
-	ginkgo.By("Creating qos policy " + eipQoSPolicyName)
-	eipRules := getEIPQoSRule(eipLimit)
-
-	eipQoSPolicy := framework.MakeQoSPolicy(eipQoSPolicyName, false, apiv1.QoSBindingTypeEIP, eipRules)
-	_ = qosPolicyClient.CreateSync(eipQoSPolicy)
-
-	ginkgo.By("Patch eip " + eipName + " with qos policy " + eipQoSPolicyName)
-	_ = eipClient.PatchQoSPolicySync(eipName, eipQoSPolicyName)
-
-	// match qos of priority 1
-	ginkgo.By("Check qos to match priority 1 is limited to " + strconv.Itoa(eipLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, eipLimit, true)
-
-	ginkgo.By("Remove qos policy " + eipQoSPolicyName + " from natgw " + eipName)
-	_ = eipClient.PatchQoSPolicySync(eipName, "")
-
-	ginkgo.By("Deleting qos policy " + eipQoSPolicyName)
-	qosPolicyClient.DeleteSync(eipQoSPolicyName)
-
-	// match qos of priority 2
-	ginkgo.By("Check qos to match priority 2 is limited to " + strconv.Itoa(specificIPLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, specificIPLimit, true)
-
-	// change qos policy of natgw
-	newNatGwQoSPolicyName := "new-priority-nic-qos-policy-" + framework.RandomSuffix()
-	ginkgo.By("Creating qos policy " + newNatGwQoSPolicyName)
-	newNatgwRules := getNicDefaultQoSPolicy(defaultNicLimit)
-
-	newNatgwQoSPolicy := framework.MakeQoSPolicy(newNatGwQoSPolicyName, true, apiv1.QoSBindingTypeNatGw, newNatgwRules)
-	_ = qosPolicyClient.CreateSync(newNatgwQoSPolicy)
-
-	ginkgo.By("Change qos policy of natgw " + natgwName + " to " + newNatGwQoSPolicyName)
-	_ = vpcNatGwClient.PatchQoSPolicySync(natgwName, newNatGwQoSPolicyName)
-
-	// match qos of priority 3
-	ginkgo.By("Check qos to match priority 3 is limited to " + strconv.Itoa(specificIPLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, defaultNicLimit, true)
-
-	ginkgo.By("Remove qos policy " + natGwQoSPolicyName + " from natgw " + natgwName)
-	_ = vpcNatGwClient.PatchQoSPolicySync(natgwName, "")
-
-	ginkgo.By("Deleting qos policy " + natGwQoSPolicyName)
-	qosPolicyClient.DeleteSync(natGwQoSPolicyName)
-
-	ginkgo.By("Deleting qos policy " + newNatGwQoSPolicyName)
-	qosPolicyClient.DeleteSync(newNatGwQoSPolicyName)
-
-	ginkgo.By("Check qos " + natGwQoSPolicyName + " is not limited to " + strconv.Itoa(defaultNicLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, defaultNicLimit, false)
-}
-
-func createNatGwAndSetQosCases(f *framework.Framework,
-	vpcNatGwClient *framework.VpcNatGatewayClient,
-	ipClient *framework.IPClient,
-	eipClient *framework.IptablesEIPClient,
-	fipClient *framework.IptablesFIPClient,
-	subnetClient *framework.SubnetClient,
-	qosPolicyClient *framework.QoSPolicyClient,
-	vpc1Pod *corev1.Pod,
-	vpc2Pod *corev1.Pod,
-	vpc2EIP *apiv1.IptablesEIP,
-	natgwName string,
-	eipName string,
-	fipName string,
-	vpcName string,
-	overlaySubnetName string,
-	lanIP string,
-	attachDefName string,
-) {
-	ginkgo.GinkgoHelper()
-
-	// delete fip
-	ginkgo.By("Deleting fip " + fipName)
-	fipClient.DeleteSync(fipName)
-
-	ginkgo.By("Deleting eip " + eipName)
-	eipClient.DeleteSync(eipName)
-
-	// the only pod for vpc nat gateway
-	vpcNatGw1PodName := util.GenNatGwPodName(natgwName)
-
-	// delete vpc nat gw statefulset remaining ip for eth0 and net2
-	ginkgo.By("Deleting custom vpc nat gw " + natgwName)
-	vpcNatGwClient.DeleteSync(natgwName)
-
-	overlaySubnet1 := subnetClient.Get(overlaySubnetName)
-	macvlanSubnet := subnetClient.Get(attachDefName)
-	eth0IpName := ovs.PodNameToPortName(vpcNatGw1PodName, framework.KubeOvnNamespace, overlaySubnet1.Spec.Provider)
-	net1IpName := ovs.PodNameToPortName(vpcNatGw1PodName, framework.KubeOvnNamespace, macvlanSubnet.Spec.Provider)
-	ginkgo.By("Deleting vpc nat gw eth0 ip " + eth0IpName)
-	ipClient.DeleteSync(eth0IpName)
-	ginkgo.By("Deleting vpc nat gw net1 ip " + net1IpName)
-	ipClient.DeleteSync(net1IpName)
-
-	natgwQoSPolicyName := "default-nic-qos-policy-" + framework.RandomSuffix()
-	ginkgo.By("Creating qos policy " + natgwQoSPolicyName)
-	rules := getNicDefaultQoSPolicy(defaultNicLimit)
-
-	qosPolicy := framework.MakeQoSPolicy(natgwQoSPolicyName, true, apiv1.QoSBindingTypeNatGw, rules)
-	_ = qosPolicyClient.CreateSync(qosPolicy)
-
-	ginkgo.By("Creating custom vpc nat gw")
-	vpcNatGw := framework.MakeVpcNatGateway(natgwName, vpcName, overlaySubnetName, lanIP, attachDefName, natgwQoSPolicyName)
-	_ = vpcNatGwClient.CreateSync(vpcNatGw, f.ClientSet)
-
-	eipQoSPolicyName := "eip-qos-policy-" + framework.RandomSuffix()
-	ginkgo.By("Creating qos policy " + eipQoSPolicyName)
-	rules = getEIPQoSRule(eipLimit)
-
-	eipQoSPolicy := framework.MakeQoSPolicy(eipQoSPolicyName, false, apiv1.QoSBindingTypeEIP, rules)
-	_ = qosPolicyClient.CreateSync(eipQoSPolicy)
-
-	ginkgo.By("Creating eip " + eipName)
-	vpc1EIP := framework.MakeIptablesEIP(eipName, "", "", "", natgwName, attachDefName, eipQoSPolicyName)
-	vpc1EIP = eipClient.CreateSync(vpc1EIP)
-
-	ginkgo.By("Creating fip " + fipName)
-	fip := framework.MakeIptablesFIPRule(fipName, eipName, vpc1Pod.Status.PodIP)
-	_ = fipClient.CreateSync(fip)
-
-	ginkgo.By("Check qos " + eipQoSPolicyName + " is limited to " + strconv.Itoa(eipLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, eipLimit, true)
-
-	ginkgo.By("Remove qos policy " + eipQoSPolicyName + " from natgw " + natgwName)
-	_ = eipClient.PatchQoSPolicySync(eipName, "")
-
-	ginkgo.By("Check qos " + natgwQoSPolicyName + " is limited to " + strconv.Itoa(defaultNicLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, defaultNicLimit, true)
-
-	ginkgo.By("Remove qos policy " + natgwQoSPolicyName + " from natgw " + natgwName)
-	_ = vpcNatGwClient.PatchQoSPolicySync(natgwName, "")
-
-	ginkgo.By("Check qos " + natgwQoSPolicyName + " is not limited to " + strconv.Itoa(defaultNicLimit) + "Mbps")
-	checkQos(f, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, defaultNicLimit, false)
-
-	ginkgo.By("Deleting qos policy " + natgwQoSPolicyName)
-	qosPolicyClient.DeleteSync(natgwQoSPolicyName)
-
-	ginkgo.By("Deleting qos policy " + eipQoSPolicyName)
-	qosPolicyClient.DeleteSync(eipQoSPolicyName)
-}
-
-func validRateLimit(text string, limit int) bool {
-	maxValue := float64(limit) * 1024 * 1024 * 1.2
-	minValue := float64(limit) * 1024 * 1024 * 0.8
-	lines := strings.SplitSeq(text, "\n")
-	for line := range lines {
-		if line == "" {
-			continue
-		}
-		fields := strings.Split(line, ",")
-		number, err := strconv.Atoi(fields[len(fields)-1])
-		if err != nil {
-			continue
-		}
-		if v := float64(number); v >= minValue && v <= maxValue {
-			return true
-		}
-	}
-	return false
-}
-
-var _ = framework.Describe("[group:qos-policy]", func() {
-	f := framework.NewDefaultFramework("qos-policy")
-
-	var skip bool
-	var cs clientset.Interface
-	var attachNetClient *framework.NetworkAttachmentDefinitionClient
-	var clusterName string
-	var vpcClient *framework.VpcClient
-	var vpcNatGwClient *framework.VpcNatGatewayClient
-	var subnetClient *framework.SubnetClient
-	var podClient *framework.PodClient
-	var ipClient *framework.IPClient
-	var iptablesEIPClient *framework.IptablesEIPClient
-	var iptablesFIPClient *framework.IptablesFIPClient
-	var qosPolicyClient *framework.QoSPolicyClient
-
-	var net1NicName string
-	var dockerExtNetName string
-
-	// docker network
-	var dockerExtNetNetwork *dockernetwork.Inspect
-
-	var vpcQosParams *qosParams
-	var vpc1Pod *corev1.Pod
-	var vpc2Pod *corev1.Pod
-	var vpc1EIP *apiv1.IptablesEIP
-	var vpc2EIP *apiv1.IptablesEIP
-	var vpc1FIP *apiv1.IptablesFIPRule
-	var vpc2FIP *apiv1.IptablesFIPRule
-
-	var lanIP string
-	var overlaySubnetV4Cidr string
-	var overlaySubnetV4Gw string
-	var eth0Exist, net1Exist bool
-	var annotations1 map[string]string
-	var annotations2 map[string]string
-	var iperfServerCmd []string
-
-	ginkgo.BeforeEach(func() {
-		vpcQosParams = newVPCQoSParamsInit()
-
-		dockerExtNetName = "kube-ovn-qos-" + framework.RandomSuffix()
-
-		vpcQosParams.vpc1SubnetName = "qos-vpc1-subnet-" + framework.RandomSuffix()
-		vpcQosParams.vpc2SubnetName = "qos-vpc2-subnet-" + framework.RandomSuffix()
-
-		vpcQosParams.vpcNat1GwName = "qos-gw1-" + framework.RandomSuffix()
-		vpcQosParams.vpcNat2GwName = "qos-gw2-" + framework.RandomSuffix()
-
-		vpcQosParams.vpc1EIPName = "qos-vpc1-eip-" + framework.RandomSuffix()
-		vpcQosParams.vpc2EIPName = "qos-vpc2-eip-" + framework.RandomSuffix()
-
-		vpcQosParams.vpc1FIPName = "qos-vpc1-fip-" + framework.RandomSuffix()
-		vpcQosParams.vpc2FIPName = "qos-vpc2-fip-" + framework.RandomSuffix()
-
-		vpcQosParams.vpc1PodName = "qos-vpc1-pod-" + framework.RandomSuffix()
-		vpcQosParams.vpc2PodName = "qos-vpc2-pod-" + framework.RandomSuffix()
-
-		vpcQosParams.attachDefName = "qos-ovn-vpc-external-network-" + framework.RandomSuffix()
-		vpcQosParams.subnetProvider = fmt.Sprintf("%s.%s", vpcQosParams.attachDefName, framework.KubeOvnNamespace)
-
-		cs = f.ClientSet
-		podClient = f.PodClient()
-		attachNetClient = f.NetworkAttachmentDefinitionClientNS(framework.KubeOvnNamespace)
-		subnetClient = f.SubnetClient()
-		vpcClient = f.VpcClient()
-		vpcNatGwClient = f.VpcNatGatewayClient()
-		iptablesEIPClient = f.IptablesEIPClient()
-		ipClient = f.IPClient()
-		iptablesFIPClient = f.IptablesFIPClient()
-		qosPolicyClient = f.QoSPolicyClient()
-
-		if skip {
-			ginkgo.Skip("underlay spec only runs on kind clusters")
-		}
-
-		if clusterName == "" {
-			ginkgo.By("Getting k8s nodes")
-			k8sNodes, err := e2enode.GetReadySchedulableNodes(context.Background(), cs)
-			framework.ExpectNoError(err)
-
-			cluster, ok := kind.IsKindProvided(k8sNodes.Items[0].Spec.ProviderID)
-			if !ok {
-				skip = true
-				ginkgo.Skip("underlay spec only runs on kind clusters")
-			}
-			clusterName = cluster
-		}
-
-		ginkgo.By("Ensuring docker network " + dockerExtNetName + " exists")
-		network, err := docker.NetworkCreate(dockerExtNetName, true, true)
-		framework.ExpectNoError(err, "creating docker network "+dockerExtNetName)
-		dockerExtNetNetwork = network
-
-		ginkgo.By("Getting kind nodes")
-		nodes, err := kind.ListNodes(clusterName, "")
-		framework.ExpectNoError(err, "getting nodes in kind cluster")
-		framework.ExpectNotEmpty(nodes)
-
-		ginkgo.By("Connecting nodes to the docker network")
-		err = kind.NetworkConnect(dockerExtNetNetwork.ID, nodes)
-		framework.ExpectNoError(err, "connecting nodes to network "+dockerExtNetName)
-
-		ginkgo.By("Getting node links that belong to the docker network")
-		nodes, err = kind.ListNodes(clusterName, "")
-		framework.ExpectNoError(err, "getting nodes in kind cluster")
-
-		ginkgo.By("Validating node links")
-		network1, err := docker.NetworkInspect(dockerExtNetName)
-		framework.ExpectNoError(err)
-		for _, node := range nodes {
-			links, err := node.ListLinks()
-			framework.ExpectNoError(err, "failed to list links on node %s: %v", node.Name(), err)
-			net1Mac := network1.Containers[node.ID].MacAddress
-			for _, link := range links {
-				ginkgo.By("exist node nic " + link.IfName)
-				if link.IfName == "eth0" {
-					eth0Exist = true
-				}
-				if link.Address == net1Mac.String() {
-					net1NicName = link.IfName
-					net1Exist = true
-				}
-			}
-			framework.ExpectTrue(eth0Exist)
-			framework.ExpectTrue(net1Exist)
-		}
-		setupNetworkAttachmentDefinition(
-			f, dockerExtNetNetwork, attachNetClient,
-			subnetClient, vpcQosParams.attachDefName, net1NicName, vpcQosParams.subnetProvider, dockerExtNetName)
-	})
-
-	ginkgo.AfterEach(func() {
-		ginkgo.By("Deleting macvlan underlay subnet " + vpcQosParams.attachDefName)
-		subnetClient.DeleteSync(vpcQosParams.attachDefName)
-
-		// delete net1 attachment definition
-		ginkgo.By("Deleting nad " + vpcQosParams.attachDefName)
-		attachNetClient.Delete(vpcQosParams.attachDefName)
-
-		ginkgo.By("Getting nodes")
-		nodes, err := kind.ListNodes(clusterName, "")
-		framework.ExpectNoError(err, "getting nodes in cluster")
-
-		if dockerExtNetNetwork != nil {
-			ginkgo.By("Disconnecting nodes from the docker network")
-			err = kind.NetworkDisconnect(dockerExtNetNetwork.ID, nodes)
-			framework.ExpectNoError(err, "disconnecting nodes from network "+dockerExtNetName)
-			ginkgo.By("Deleting docker network " + dockerExtNetName + " exists")
-			err := docker.NetworkRemove(dockerExtNetNetwork.ID)
-			framework.ExpectNoError(err, "deleting docker network "+dockerExtNetName)
-		}
-	})
-
-	_ = framework.Describe("vpc qos", func() {
-		ginkgo.BeforeEach(func() {
-			iperfServerCmd = []string{"iperf", "-s", "-i", "1", "-p", iperf2Port}
-			overlaySubnetV4Cidr = "10.0.0.0/24"
-			overlaySubnetV4Gw = "10.0.0.1"
-			lanIP = "10.0.0.254"
-			natgwQoS := ""
-			setupVpcNatGwTestEnvironment(
-				f, dockerExtNetNetwork, attachNetClient,
-				subnetClient, vpcClient, vpcNatGwClient,
-				vpcQosParams.vpc1Name, vpcQosParams.vpc1SubnetName, vpcQosParams.vpcNat1GwName,
-				natgwQoS, overlaySubnetV4Cidr, overlaySubnetV4Gw, lanIP,
-				dockerExtNetName, vpcQosParams.attachDefName, net1NicName,
-				vpcQosParams.subnetProvider,
-				true,
-			)
-			annotations1 = map[string]string{
-				util.LogicalSwitchAnnotation: vpcQosParams.vpc1SubnetName,
-			}
-			ginkgo.By("Creating pod " + vpcQosParams.vpc1PodName)
-			vpc1Pod = framework.MakePod(f.Namespace.Name, vpcQosParams.vpc1PodName, nil, annotations1, framework.AgnhostImage, iperfServerCmd, nil)
-			vpc1Pod = podClient.CreateSync(vpc1Pod)
-
-			ginkgo.By("Creating eip " + vpcQosParams.vpc1EIPName)
-			vpc1EIP = framework.MakeIptablesEIP(vpcQosParams.vpc1EIPName, "", "", "", vpcQosParams.vpcNat1GwName, vpcQosParams.attachDefName, "")
-			vpc1EIP = iptablesEIPClient.CreateSync(vpc1EIP)
-
-			ginkgo.By("Creating fip " + vpcQosParams.vpc1FIPName)
-			vpc1FIP = framework.MakeIptablesFIPRule(vpcQosParams.vpc1FIPName, vpcQosParams.vpc1EIPName, vpc1Pod.Status.PodIP)
-			_ = iptablesFIPClient.CreateSync(vpc1FIP)
-
-			setupVpcNatGwTestEnvironment(
-				f, dockerExtNetNetwork, attachNetClient,
-				subnetClient, vpcClient, vpcNatGwClient,
-				vpcQosParams.vpc2Name, vpcQosParams.vpc2SubnetName, vpcQosParams.vpcNat2GwName,
-				natgwQoS, overlaySubnetV4Cidr, overlaySubnetV4Gw, lanIP,
-				dockerExtNetName, vpcQosParams.attachDefName, net1NicName,
-				vpcQosParams.subnetProvider,
-				true,
-			)
-
-			annotations2 = map[string]string{
-				util.LogicalSwitchAnnotation: vpcQosParams.vpc2SubnetName,
-			}
-
-			ginkgo.By("Creating pod " + vpcQosParams.vpc2PodName)
-			vpc2Pod = framework.MakePod(f.Namespace.Name, vpcQosParams.vpc2PodName, nil, annotations2, framework.AgnhostImage, iperfServerCmd, nil)
-			vpc2Pod = podClient.CreateSync(vpc2Pod)
-
-			ginkgo.By("Creating eip " + vpcQosParams.vpc2EIPName)
-			vpc2EIP = framework.MakeIptablesEIP(vpcQosParams.vpc2EIPName, "", "", "", vpcQosParams.vpcNat2GwName, vpcQosParams.attachDefName, "")
-			vpc2EIP = iptablesEIPClient.CreateSync(vpc2EIP)
-
-			ginkgo.By("Creating fip " + vpcQosParams.vpc2FIPName)
-			vpc2FIP = framework.MakeIptablesFIPRule(vpcQosParams.vpc2FIPName, vpcQosParams.vpc2EIPName, vpc2Pod.Status.PodIP)
-			_ = iptablesFIPClient.CreateSync(vpc2FIP)
-		})
-		ginkgo.AfterEach(func() {
-			ginkgo.By("Deleting fip " + vpcQosParams.vpc1FIPName)
-			iptablesFIPClient.DeleteSync(vpcQosParams.vpc1FIPName)
-
-			ginkgo.By("Deleting fip " + vpcQosParams.vpc2FIPName)
-			iptablesFIPClient.DeleteSync(vpcQosParams.vpc2FIPName)
-
-			ginkgo.By("Deleting eip " + vpcQosParams.vpc1EIPName)
-			iptablesEIPClient.DeleteSync(vpcQosParams.vpc1EIPName)
-
-			ginkgo.By("Deleting eip " + vpcQosParams.vpc2EIPName)
-			iptablesEIPClient.DeleteSync(vpcQosParams.vpc2EIPName)
-
-			ginkgo.By("Deleting pod " + vpcQosParams.vpc1PodName)
-			podClient.DeleteSync(vpcQosParams.vpc1PodName)
-
-			ginkgo.By("Deleting pod " + vpcQosParams.vpc2PodName)
-			podClient.DeleteSync(vpcQosParams.vpc2PodName)
-
-			ginkgo.By("Deleting custom vpc nat gw " + vpcQosParams.vpcNat1GwName)
-			vpcNatGwClient.DeleteSync(vpcQosParams.vpcNat1GwName)
-
-			ginkgo.By("Deleting custom vpc nat gw " + vpcQosParams.vpcNat2GwName)
-			vpcNatGwClient.DeleteSync(vpcQosParams.vpcNat2GwName)
-
-			// the only pod for vpc nat gateway
-			vpcNatGw1PodName := util.GenNatGwPodName(vpcQosParams.vpcNat1GwName)
-
-			// delete vpc nat gw statefulset remaining ip for eth0 and net2
-			overlaySubnet1 := subnetClient.Get(vpcQosParams.vpc1SubnetName)
-			macvlanSubnet := subnetClient.Get(vpcQosParams.attachDefName)
-			eth0IpName := ovs.PodNameToPortName(vpcNatGw1PodName, framework.KubeOvnNamespace, overlaySubnet1.Spec.Provider)
-			net1IpName := ovs.PodNameToPortName(vpcNatGw1PodName, framework.KubeOvnNamespace, macvlanSubnet.Spec.Provider)
-			ginkgo.By("Deleting vpc nat gw eth0 ip " + eth0IpName)
-			ipClient.DeleteSync(eth0IpName)
-			ginkgo.By("Deleting vpc nat gw net1 ip " + net1IpName)
-			ipClient.DeleteSync(net1IpName)
-			ginkgo.By("Deleting overlay subnet " + vpcQosParams.vpc1SubnetName)
-			subnetClient.DeleteSync(vpcQosParams.vpc1SubnetName)
-
-			ginkgo.By("Getting overlay subnet " + vpcQosParams.vpc2SubnetName)
-			overlaySubnet2 := subnetClient.Get(vpcQosParams.vpc2SubnetName)
-
-			vpcNatGw2PodName := util.GenNatGwPodName(vpcQosParams.vpcNat2GwName)
-			eth0IpName = ovs.PodNameToPortName(vpcNatGw2PodName, framework.KubeOvnNamespace, overlaySubnet2.Spec.Provider)
-			net1IpName = ovs.PodNameToPortName(vpcNatGw2PodName, framework.KubeOvnNamespace, macvlanSubnet.Spec.Provider)
-			ginkgo.By("Deleting vpc nat gw eth0 ip " + eth0IpName)
-			ipClient.DeleteSync(eth0IpName)
-			ginkgo.By("Deleting vpc nat gw net1 ip " + net1IpName)
-			ipClient.DeleteSync(net1IpName)
-			ginkgo.By("Deleting overlay subnet " + vpcQosParams.vpc2SubnetName)
-			subnetClient.DeleteSync(vpcQosParams.vpc2SubnetName)
-
-			ginkgo.By("Deleting custom vpc " + vpcQosParams.vpc1Name)
-			vpcClient.DeleteSync(vpcQosParams.vpc1Name)
-
-			ginkgo.By("Deleting custom vpc " + vpcQosParams.vpc2Name)
-			vpcClient.DeleteSync(vpcQosParams.vpc2Name)
-		})
-		framework.ConformanceIt("default nic qos", func() {
-			// case 1: set qos policy for natgw
-			// case 2: rebuild qos when natgw pod restart
-			defaultQoSCases(f, vpcNatGwClient, podClient, qosPolicyClient, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, vpcQosParams.vpcNat1GwName)
-		})
-		framework.ConformanceIt("eip qos", func() {
-			// case 1: set qos policy for eip
-			// case 2: update qos policy for eip
-			// case 3: change qos policy of eip
-			// case 4: rebuild qos when natgw pod restart
-			eipQoSCases(f, iptablesEIPClient, podClient, qosPolicyClient, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, vpcQosParams.vpc1EIPName, vpcQosParams.vpcNat1GwName)
-		})
-		framework.ConformanceIt("specifying ip qos", func() {
-			// case 1: set specific ip qos policy for natgw
-			specifyingIPQoSCases(f, vpcNatGwClient, qosPolicyClient, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, vpcQosParams.vpcNat1GwName)
-		})
-		framework.ConformanceIt("qos priority matching", func() {
-			// case 1: test qos match priority
-			// case 2: change qos policy of natgw
-			priorityQoSCases(f, vpcNatGwClient, iptablesEIPClient, qosPolicyClient, vpc1Pod, vpc2Pod, vpc1EIP, vpc2EIP, vpcQosParams.vpcNat1GwName, vpcQosParams.vpc1EIPName)
-		})
-		framework.ConformanceIt("create resource with qos policy", func() {
-			// case 1: test qos when create natgw with qos policy
-			// case 2: test qos when create eip with qos policy
-			createNatGwAndSetQosCases(f,
-				vpcNatGwClient, ipClient, iptablesEIPClient, iptablesFIPClient,
-				subnetClient, qosPolicyClient, vpc1Pod, vpc2Pod, vpc2EIP, vpcQosParams.vpcNat1GwName,
-				vpcQosParams.vpc1EIPName, vpcQosParams.vpc1FIPName, vpcQosParams.vpc1Name,
-				vpcQosParams.vpc1SubnetName, lanIP, vpcQosParams.attachDefName)
-		})
 	})
 })
